Pull sparse matrix stuff into base (not including suitesparse.jl). Introduce the Sparse module. Merge linalg_sparse.jl into sparse.jl. Sparse solvers continue to remain in extras/suitesparse.jl.
